Object Recognation app with Google Cloud and Android/python-vision/requirements.txt
google-cloud-vision==0.41.0
Flask==1.0.2
gunicorn==19.6.0
google-api-python-client==1.7.7
GoogleAppEngineCloudStorageClient==1.9.22.1